requirements.txt
c8y-api>=1.9.0
pulsar-client>=3.4.0
pandas>=2.3.3